pushing all changes

+++ whisper-wave/src/app/components/registerForm.tsx
@@ -3,10 +3,12 @@ import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { useState } from "react";
+import "./../styles.css";
 
 const signInForm = (displayStatus: string) => {
   const [userName, setUserNameState] = useState("");
   const [passwordName, setPasswordState] = useState("");
+  const [confirmPasswordName, setConfirmPasswordState] = useState("");
   return (
     <>
       <div
@@ -15,20 +17,27 @@ const signInForm = (displayStatus: string) => {
         <Form
           onSubmit={async (e) => {
             e.preventDefault();
-            const response = await fetch(`http://127.0.0.1:8000/signInUser`, {
-              method: "POST",
-              headers: {
-                "Content-Type": "application/json", // Add this line
-              },
-              body: JSON.stringify({
-                username: `${userName}`,
-                password: `${passwordName}`,
-              }),
-            }).then((response) => response.json());
-            if ((response.message = "user successfully signed in")) {
-              localStorage.setItem("username", `${userName}`);
-              console.log("line 27, user successfully signed in");
-              location.reload();
+            if (confirmPasswordName === passwordName) {
+              const response = await fetch(
+                `http://127.0.0.1:8000/registerUser`,
+                {
+                  method: "POST",
+                  headers: {
+                    "Content-Type": "application/json", // Add this line
+                  },
+                  body: JSON.stringify({
+                    username: `${userName}`,
+                    password: `${passwordName}`,
+                  }),
+                }
+              ).then((response) => response.json());
+              if ((response.message = "user successfully signed in")) {
+                localStorage.setItem("username", `${userName}`);
+                console.log("line 27, user successfully signed in");
+                location.reload();
+              }
+            } else {
+              alert("Password confirmation and password do not match");
             }
           }}
         >
@@ -51,6 +60,19 @@ const signInForm = (displayStatus: string) => {
               placeholder="Password"
               onChange={async (e) => {
                 setPasswordState(e.target.value);
+              }}
+            />
+          </FloatingLabel>
+          <FloatingLabel
+            className="mt-2"
+            controlId="floatingPassword"
+            label="Confirm Password"
+          >
+            <Form.Control
+              type="password"
+              placeholder="Confirm Password"
+              onChange={async (e) => {
+                setConfirmPasswordState(e.target.value);
               }}
             />
           </FloatingLabel>
